% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_stubs.R
\name{fill_stubs}
\alias{fill_stubs}
\title{Fill Stubs}
\usage{
fill_stubs(x)
}
\arguments{
\item{x}{variable - missing values will be filled with the last non-NA value}
}
\value{
vector of the same type as x with missing values filled in
}
\description{
Variable x is assumed to consist of stubs (assuming the data stems from converting a table with stubs, such as pivot tables in Excel).
}
