% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_code.R
\name{replace_code}
\alias{replace_code}
\title{Re-code vector x based on codebook}
\usage{
replace_code(x, translate, xname = NULL, as.factor = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{translate}{data frame containing a code book. It is assumed that it has
columns called \code{Variable}, \code{Value}, and \code{Label}}

\item{xname}{name of the variable of \code{x} used in the codebook. By default,
the name of the variable is used for a search, regular expressions in xname
can be used to resolve disambiguity.}

\item{as.factor}{binary variable - should the result be returned as factor?
In that case, factor levels are ordered according to the original values in x.}
}
\value{
vector of the same length as x with re-coded values
}
\description{
Re-code vector x based on codebook
}
