% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_vals.R
\name{replace_vals}
\alias{replace_vals}
\title{Replace values in vector}
\usage{
replace_vals(x, values, labels)
}
\arguments{
\item{x}{vector of values to be replaced}

\item{values}{vector of values in x to be replaced.}

\item{labels}{vector of the same length as \code{values}. It is assumed that values
are replaced 1-1, i.e. the ith value in values is replaced by the ith value in labels}
}
\description{
The \code{replace_vals} function allows re-coding a set of values in vector \code{x}
(in most cases a vector of integers) by their corresponding \code{labels}.
}
